Coverity!

+++ rpcs3/Emu/Cell/PPCDecoder.cpp
@@ -2,7 +2,7 @@
 #include "Emu/Memory/Memory.h"
 #include "PPCDecoder.h"
 
-u8 PPCDecoder::DecodeMemory(const u32 address)
+u32 PPCDecoder::DecodeMemory(const u32 address)
 {
 	u32 instr = vm::read32(address);
 	Decode(instr);
